alignment scripts containing additions made by Felix
These alignment scripts contain the option to use a manual shift to check the spectral index map. The 'alignImages_FMP.py' runs the 'align_imagesEHTim_FMP.py' file with fits image files and input parameters for masking provided.
The script now also includes an option to calculate both statistical and systematic errors due to image alignment via Monte-Carlo simulations. This can be turned on or off with the do_err keywordarg. Details are described in the comments. This is still work in progress, so any suggestions are very welcome.